requirements.txt
streamlit>=1.32.0
openai>=1.43.0
pydantic>=2.8.0
python-dotenv>=1.0.1
pandas>=2.2.0
numpy>=1.26.0
google-generativeai>=0.7.2
flask>=3.0.0
twilio>=9.0.0
requests>=2.31.0

